backend/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
flask-cors==4.0.0
openai==1.3.0
google-generativeai==0.8.3
anthropic==0.7.0
protobuf>=5.29.1
